refactor : Apple Music SearchBaseResult 추가
* 검색 API의 경우 데이터 형식이 동일하여 공통 BaseResult 추가
